predelana a opravena sprava nametu
git-svn-id: https://zdrojaky.abclinuxu.cz/svn-abclinuxu/trunk@3292 be82d567-0701-4079-a524-1ea639e27e29

+++ src/cz/abclinuxu/utils/email/EmailSender.java
@@ -229,7 +229,7 @@ public class EmailSender implements Configurable {
 
             for ( Iterator iter = users.iterator(); iter.hasNext(); ) {
                 try {
-                    user.setId(((Integer) iter.next()).intValue());
+                    user.setId((Integer) iter.next());
                     user = (User) persistence.findById(user);
 
                     // check, that user has valid email
